+++ src/category.rs
@@ -1,52 +1,76 @@
-use serde::Deserialize;
+use std::cmp::Ordering;
 
-use super::{Database, Error, Type};
+use serde::{Deserialize, Serialize};
+
+use crate::{Indexable, IndexedMapUpdate};
+
+use super::{Database, Error};
 
 /// A category for a cluster.
-#[derive(Deserialize, Queryable)]
+#[derive(Debug, Deserialize, Queryable, Serialize, PartialEq, Eq)]
 pub struct Category {
-    pub id: i32,
+    pub id: u32,
     pub name: String,
 }
 
+impl PartialOrd for Category {
+    fn partial_cmp(&self, other: &Self) -> Option<Ordering> {
+        Some(self.cmp(other))
+    }
+}
+
+impl Ord for Category {
+    fn cmp(&self, other: &Self) -> Ordering {
+        let ord = self.name.cmp(&other.name);
+        match ord {
+            Ordering::Equal => self.id.cmp(&other.id),
+            _ => ord,
+        }
+    }
+}
+
+impl Indexable for Category {
+    fn key(&self) -> &[u8] {
+        self.name.as_bytes()
+    }
+
+    fn value(&self) -> Vec<u8> {
+        use bincode::Options;
+
+        bincode::DefaultOptions::new()
+            .serialize(self)
+            .expect("serializable")
+    }
+
+    fn set_index(&mut self, index: u32) {
+        self.id = index;
+    }
+}
+
+impl IndexedMapUpdate for Category {
+    type Entry = Category;
+
+    fn key(&self) -> Option<&[u8]> {
+        if self.name.is_empty() {
+            None
+        } else {
+            Some(self.name.as_bytes())
+        }
+    }
+
+    fn apply(&self, mut value: Self::Entry) -> Result<Self::Entry, anyhow::Error> {
+        value.name.clear();
+        value.name.push_str(&self.name);
+
+        Ok(value)
+    }
+
+    fn verify(&self, value: &Self::Entry) -> bool {
+        self.name == value.name
+    }
+}
+
 impl Database {
-    /// Adds a new category to the database.
-    ///
-    /// # Errors
-    ///
-    /// Returns an error if the database operation fails.
-    pub async fn add_category(&self, name: &str) -> Result<i32, Error> {
-        let conn = self.pool.get().await?;
-        conn.insert_into("category", &[("name", Type::TEXT)], &[&name])
-            .await
-    }
-
-    /// Returns the number of categories in the database.
-    ///
-    /// # Errors
-    ///
-    /// Returns an error if the database query fails.
-    pub async fn count_categories(&self) -> Result<i64, Error> {
-        let conn = self.pool.get().await?;
-        conn.count("category", &[], &[], &[]).await
-    }
-
-    /// Returns the category with the given ID.
-    ///
-    /// # Errors
-    ///
-    /// Returns an error if the database query fails.
-    pub async fn load_category(&self, id: i32) -> Result<Category, Error> {
-        let conn = self.pool.get().await?;
-        conn.select_one_from::<Category>(
-            "category",
-            &["id", "name"],
-            &[("id", super::Type::INT4)],
-            &[&id],
-        )
-        .await
-    }
-
     /// Returns the categories between `after` and `before`.
     ///
     /// # Errors
@@ -94,22 +118,23 @@ impl Database {
         }
 
         let mut conn = self.pool.get_diesel_conn().await?;
-        let mut rows: Vec<Category> = query.get_results(&mut conn).await?;
-        if !is_first {
-            rows = rows.into_iter().rev().collect();
-        }
+        let rows: Vec<(i32, String)> = query.get_results(&mut conn).await?;
+        let rows = if is_first {
+            rows.into_iter()
+                .map(|(id, name)| {
+                    let id = u32::from_ne_bytes(id.to_ne_bytes());
+                    Category { id, name }
+                })
+                .collect()
+        } else {
+            rows.into_iter()
+                .rev()
+                .map(|(id, name)| {
+                    let id = u32::from_ne_bytes(id.to_ne_bytes());
+                    Category { id, name }
+                })
+                .collect()
+        };
         Ok(rows)
-    }
-
-    /// Updates the category with the given ID.
-    ///
-    /// # Errors
-    ///
-    /// Returns an error if the database query fails.
-    pub async fn update_category(&self, id: i32, name: &str) -> Result<(), Error> {
-        let conn = self.pool.get().await?;
-        conn.update("category", id, &[("name", Type::TEXT)], &[&name])
-            .await?;
-        Ok(())
     }
 }