+++ src/migration.rs
@@ -34,6 +34,102 @@ use tracing::info;
 /// ```
 const COMPATIBLE_VERSION_REQ: &str = ">=0.22.0,<0.23.0-alpha";
 
+/// Migrates data exists in `PostgresQL` to Rocksdb if necessary.
+///
+/// Migration is supported for current released version only. And the migrated data
+/// and related interface should be removed from `PostgresQL` database in the next released
+/// version.
+///
+/// # Errors
+///
+/// Returns an error if the data hasn't been migrated successfully to Rocksdb.
+pub async fn migrate_backend<P: AsRef<Path>>(
+    db: &super::Database,
+    store: &super::Store,
+    data_dir: P,
+) -> Result<()> {
+    let path = data_dir.as_ref();
+    let file = path.join("VERSION");
+
+    let version = read_version_file(&file)?;
+
+    if VersionReq::parse(COMPATIBLE_VERSION_REQ)?.matches(&version) {
+        backend_0_22(db, store).await?;
+    }
+    Ok(())
+}
+
+async fn backend_0_22(db: &super::Database, store: &super::Store) -> Result<()> {
+    tracing::info!("starting to transfer category data...");
+    let mut after = None;
+    let mut categories = vec![];
+
+    // retrieve categories from Postgres
+    while let Ok(cats) = db.load_categories(&after, &None, true, 100).await {
+        if cats.is_empty() {
+            break;
+        }
+        after = cats.last().map(|c| {
+            (
+                i32::try_from(c.id).expect("id out of range"),
+                c.name.clone(),
+            )
+        });
+        categories.extend(cats);
+    }
+
+    let mut table = store.category_map();
+
+    if table.count()? > 0 {
+        tracing::info!("category data migration completed.");
+        return Ok(());
+    }
+
+    let max_id = categories
+        .iter()
+        .map(|c| c.id)
+        .max()
+        .expect("maximum id doesn't exist");
+
+    for id in 0..=max_id {
+        if let Ok(_c) = table.get(id) {
+            continue;
+        }
+
+        let added = table.add(&format!("dummy{id}"))?;
+        if added != id {
+            return Err(anyhow!(
+                "corrupted category table: inserting {id} and asigned with {added}"
+            ));
+        }
+    }
+
+    for cat in &categories {
+        table.update(cat.id, &format!("dummy{}", cat.id), &cat.name)?;
+    }
+
+    if categories.len() != usize::try_from(max_id + 1).expect("max id out of range") {
+        let mut flags = vec![false; usize::try_from(max_id + 1).expect("max_id out of range")];
+        for c in &categories {
+            flags[usize::try_from(c.id).expect("max_id out of range")] = true;
+        }
+
+        for id in flags
+            .into_iter()
+            .enumerate()
+            .filter_map(|(id, exists)| if exists { None } else { Some(id) })
+        {
+            table.deactivate(u32::try_from(id).expect("id out of range"))?;
+        }
+    }
+
+    tracing::info!(
+        "{} categories are migrated with {max_id} as largest id",
+        categories.len()
+    );
+    Ok(())
+}
+
 /// Migrates the data directory to the up-to-date format if necessary.
 ///
 /// Migration is supported between released versions only. The prelease versions (alpha, beta,