+++ CHANGELOG.md
@@ -7,14 +7,28 @@ Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [0.22.0] - 2023-01-09
 
+### Added
+
+- `migrate_backend` function is provided for user to transfer data between
+  PostgreSQL and RocksDB for a seamless backend transition.
+
 ### Changed
 
 - Ensures that when updating elements in `Map` and `IndexedMap`, the system now
   checks whether the new key already exists in the database. This prevents
   unintentional overwrites or conflicts, providing a more robust and reliable
   update mechanism.
+- Moved the category table from the PostgreSQL database to RocksDB.
+  - The category table data is now stored in RocksDB for improved performance
+    and scalability.
+  - A migration function has been provided to seamlessly transition data from
+    the old PostgreSQL table to RocksDB.
 - `nodes` table's fields are modified. Migration of data is supported by function
   `migrate_0_20_to_0_22`.
+
+### Deprecated
+
+- `category` table from PostgreSQL database is now deprecated.
 
 ## [0.21.0] - 2023-12-01